% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learncurves.R
\name{learncurves}
\alias{learncurves}
\alias{linear}
\alias{hyperb}
\alias{sigmoid}
\title{Learning curve functions}
\usage{
linear(m = (9/50), yint = 1, x)

hyperb(b = NULL, m = NULL, yint = 1, x)

sigmoid(b = NULL, m = NULL, yint = 1, x)
}
\arguments{
\item{yint}{The accuracy (\code{a}) of classification or identification at time
\code{x = 0} (i.e., the y-intercept of the learning curve); default value
\code{yint = 1}.}

\item{x}{The amount of time spent learning.}

\item{b, m}{Parameters defining the shape of the learning curves; if manually
specified, overrides the default parameters or those set by specifying
\code{hyp} or \code{sig}.}
}
\value{
The accuracy (\code{a}) of classification or identification at a
given time \code{x}.
}
\description{
For a given amount of time spent learning (\code{x}), determines
classification accuracy \code{a} under different learning curves.
}
\note{
The learning curves used are as follows: \describe{
\item{linear}{\code{a(x) = mlin * x + yint}} \item{hyperbolic}{\code{a(x) =
  ((bhyp * x)/(x + mhyp)) + yint}} \item{sigmoid}{\code{a(x) = b/(1 + (((b/yint) -
  1) * exp(-m * x)))}} }
}
\examples{
linear(x = 3)
hyperb(b = 9.25, m = 1, x = 1)
sigmoid(b = 10.75, m = 0.1, x = 5)

}
