% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opttime.R
\name{opttime}
\alias{opttime}
\alias{optlin}
\alias{opthyp}
\alias{optsig}
\title{Optimal time to spend learning}
\usage{
optlin(B, m = (9/50), yint = 1, q = 1, x)

opthyp(B, hyp = NULL, b = NULL, m = NULL, yint = 1, q = 1, x)

optsig(B, sig = NULL, b = NULL, m = NULL, yint = 1, q = 1, x)
}
\arguments{
\item{B}{The threshold false positive rate \eqn{\beta}.}

\item{yint}{The accuracy (\code{a}) of classification or identification at time
\code{t = 0} (i.e., the y-intercept of the learning curve); default value
\code{yint = 1}.}

\item{q}{A value indicating whether optimal time should be calculated based
on a (positive) loss model (\code{q = 1}) or a (negative) growth model
(\code{q = -1}).}

\item{x}{The rate of loss (if \code{q = 1}) or growth (if \code{q = -1}).}

\item{hyp}{The type of hyperbolic learning curve to be used with
\code{prophyp()}.For \code{hyp = 1}, sets \code{bhyp = 9.25} and \code{mhyp
  = 1}; for \code{hyp = 2}, sets \code{bhyp = 10} and \code{mhyp = 5}.}

\item{b, m}{Parameters defining the shape of the learning curves; if
manually specified, overrides the default parameters or those set by specifying
\code{hyp} or \code{sig}.}

\item{sig}{The type of sigmoid learning curve to be used with
\code{propsig()}. For \code{sig = 1}, sets \code{b = 10} and \code{m =
  0.15}; for \code{sig = 2}, sets \code{b = 10.75} and \code{m = 0.1}.}
}
\value{
The optimal amount of time to spend learning that maximizes the
proportion correctly identified.
}
\description{
For a given rate of loss or growth \code{x}, determines the optimal time to
spend learning under different learning curves and threshold false positive
rates (\eqn{\beta}).
}
\examples{
optlin(B = 0.5, x = 0.001)
opthyp(B = 0.2, hyp = 1, x = 0.005)
optsig(B = 0.5, b = 10, m = 0.15, x = 0.002)

}
