% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propcorrect.R
\name{propcorrect}
\alias{propcorrect}
\alias{proplin}
\alias{prophyp}
\alias{propsig}
\title{Proportion of habitat correctly identified}
\usage{
proplin(B, m = (9/50), yint = 1, q = 1, x)

prophyp(B, hyp = NULL, b = NULL, m = NULL, yint = 1, q = 1, x)

propsig(B, sig = NULL, b = NULL, m = NULL, yint = 1, q = 1, x)
}
\arguments{
\item{B}{The threshold false positive rate \eqn{\beta}.}

\item{yint}{The accuracy (\code{a}) of classification or identification at time
\code{t = 0} (i.e., the y-intercept of the learning curve); default value
\code{yint = 1}.}

\item{q}{A value indicating whether optimal time should be calculated based
on loss model (\code{q = 1}) or growth model (\code{q = -1}).}

\item{x}{TThe rate of loss (if \code{q = 1}) or growth (if \code{q = -1}).}

\item{hyp}{The type of hyperbolic learning curve to be used with
\code{prophyp()}.For \code{hyp = 1}, sets \code{bhyp = 9.25} and \code{mhyp
  = 1}; for \code{hyp = 2}, sets \code{bhyp = 10} and \code{mhyp = 5}.}

\item{b, m}{Parameters defining the shape of the learning curves; if
manually specified, overrides the default parameters of those set by specifying
\code{hyp} or \code{sig}.}

\item{sig}{The type of sigmoid learning curve to be used with
\code{propsig()}. For \code{sig = 1}, sets \code{b = 10} and \code{m =
  0.15}; for \code{sig = 2}, sets \code{b = 10.75} and \code{m = 0.1}.}
}
\value{
The proportion correctly identified at the optimal time given the
rate of loss (or growth) \code{x}.
}
\description{
For a given rate of loss or growth \code{x}, determines the proportion of
habitat correctly identified at the optimal time under different learning
curves and threshold false positive rates (\eqn{\beta}).
}
\examples{
proplin(B = 0.2, x = 5)
prophyp(B = 0.5, hyp = 2, x = 1)
propsig(B = 0.2, b = 9, m = 0.2, x = 0.5)

}
