% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opttime.R
\name{opttime}
\alias{opttime}
\alias{optlin}
\alias{opthyp}
\alias{optsig}
\title{Optimal time to spend learning}
\usage{
optlin(B, m = 9/50, yint = 1, q = 1, x)

opthyp(B, b = 9.25, m = 1, yint = 1, q = 1, x)

optsig(B, b = 10, m = 0.15, yint = 1, q = 1, x)
}
\arguments{
\item{B}{The threshold false positive rate \eqn{\beta}.}

\item{yint}{The accuracy (\code{a}) of classification or identification at time
\code{t = 0} (i.e., the y-intercept of the learning curve); default value
\code{yint = 1}.}

\item{q}{A value indicating whether optimal time should be calculated based
on a (positive) loss model (\code{q = 1}) or a (negative) growth model
(\code{q = -1}).}

\item{x}{The rate of loss (if \code{q = 1}) or growth (if \code{q = -1}).}

\item{b, m}{Parameters defining the shape of the learning curves; if
manually specified, overrides the default parameters.}
}
\value{
The optimal amount of time to spend learning that maximizes the
proportion correctly identified.
}
\description{
For a given rate of loss or growth \code{x}, determines the optimal time to
spend learning under different learning curves and threshold false positive
rates (\eqn{\beta}).
}
\examples{
optlin(B = 0.5, x = 0.001)
opthyp(B = 0.2, hyp = 1, x = 0.005)
optsig(B = 0.5, b = 10, m = 0.15, x = 0.002)

}
