% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propcorrect.R
\name{propcorrect}
\alias{propcorrect}
\alias{proplin}
\alias{prophyp}
\alias{propsig}
\title{Proportion of habitat correctly identified}
\usage{
proplin(B, m, a, t, q = 1, x)

prophyp(B, b, m, a, t, q = 1, x)

propsig(B, b, m, a, t, q = 1, x)
}
\arguments{
\item{B}{The threshold false positive rate \eqn{\beta}.}

\item{a}{The accuracy of identification at time t based on the learning curve used.}

\item{t}{The amount of time to spend learning.}

\item{q}{A value indicating whether optimal time should be calculated based
on loss model (\code{q = 1}) or growth model (\code{q = -1}).}

\item{x}{A positive value for the rate of loss (if \code{q = 1}) or growth (if \code{q = -1}).}

\item{b, m}{Parameters defining the shape of the learning curves; if
manually specified, overrides the default parameters of those set by specifying
\code{hyp} or \code{sig}.}
}
\value{
The proportion correctly identified at the optimal time given the
rate of loss (or growth) \code{x}.
}
\description{
For a given rate of loss or growth \code{x}, determines the proportion of
habitat correctly identified at the optimal time under different learning
curves and threshold false positive rates (\eqn{\beta}).
}
\examples{
proplin(B = 0.2, m = 0.2, a = 4, t = 3, x = 0.1)
prophyp(B = 0.5, b = 10, m = 5, a = 2, t = 5, x = 0.01)
propsig(B = 0.2, b = 9, m = 0.2, a = 6, t = 2, x = 0.5)

}
